+++ kroxylicious-runtime/src/main/java/io/kroxylicious/proxy/internal/KafkaProxyExceptionMapper.java
@@ -226,15 +226,19 @@ public class KafkaProxyExceptionMapper {
         return errorResponse(frame, error).data();
     }
 
+    private static ListConfigResourcesResponse newListConfigResourcesV0ErrorResponse(Errors apiException) {
+        return new ListConfigResourcesResponse(new ListConfigResourcesResponseData()
+                .setErrorCode(apiException.code())
+                .setConfigResources(List.of(new ListConfigResourcesResponseData.ConfigResource().setResourceType((byte) 16))));
+    }
+
     public static AbstractResponse errorResponseForMessage(RequestHeaderData requestHeaders, ApiMessage message, ApiException apiException) {
         final short apiKey = message.apiKey();
         // Our ListConfigResourcesRequestData is deserialized off the wire and so will only have the v0 fields populated.
         // we can't just all errorResponse, which in turn uses a ListConfigResourcesRequest.Builder,
         // because that builder applies some validation that's inappropriate for our purposes.
         if (apiKey == ApiKeys.LIST_CONFIG_RESOURCES.id && requestHeaders.requestApiVersion() == 0) {
-            return new ListConfigResourcesResponse(new ListConfigResourcesResponseData()
-                    .setErrorCode(Errors.forException(apiException).code())
-                    .setConfigResources(List.of(new ListConfigResourcesResponseData.ConfigResource().setResourceType((byte) 16))));
+            return newListConfigResourcesV0ErrorResponse(Errors.forException(apiException));
         }
         return errorResponse(ApiKeys.forId(apiKey), message, requestHeaders.requestApiVersion()).getErrorResponse(apiException);
     }
@@ -245,9 +249,7 @@ public class KafkaProxyExceptionMapper {
         short apiVersion = frame.apiVersion();
         final ApiKeys apiKey = frame.apiKey();
         if (frame.apiKey() == ApiKeys.LIST_CONFIG_RESOURCES && frame.apiVersion() == 0) {
-            return new ListConfigResourcesResponse(new ListConfigResourcesResponseData()
-                    .setErrorCode(Errors.forException(error).code())
-                    .setConfigResources(List.of(new ListConfigResourcesResponseData.ConfigResource().setResourceType((byte) 16))));
+            return newListConfigResourcesV0ErrorResponse(Errors.forException(error));
         }
         final AbstractRequest req = errorResponse(apiKey, reqBody, apiVersion);
         return req.getErrorResponse(error);
