+++ kroxylicious-runtime/src/main/java/io/kroxylicious/proxy/internal/subject/DefaultTransportSubjectBuilderService.java
@@ -7,7 +7,6 @@
 package io.kroxylicious.proxy.internal.subject;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.ServiceLoader;
 import java.util.function.Function;
 import java.util.stream.Collectors;
@@ -59,7 +58,7 @@ public class DefaultTransportSubjectBuilderService implements TransportSubjectBu
             for (PrincipalAdderConf adder : addPrincipals) {
                 // call methods for validation side-effect
                 buildExtractor(adder.from());
-                buildMappingRules(adder.map());
+                MappingRule.buildMappingRules(adder.map());
                 buildPrincipalFactory(adder.principalFactory());
             }
         }
@@ -72,7 +71,7 @@ public class DefaultTransportSubjectBuilderService implements TransportSubjectBu
     public void initialize(@Nullable Config config) {
         adders = Plugins.requireConfig(this, config).addPrincipals().stream()
                 .map(addConf -> new PrincipalAdder(buildExtractor(addConf.from()),
-                        buildMappingRules(addConf.map()),
+                        MappingRule.buildMappingRules(addConf.map()),
                         buildPrincipalFactory(addConf.principalFactory())))
                 .toList();
     }
@@ -83,42 +82,6 @@ public class DefaultTransportSubjectBuilderService implements TransportSubjectBu
                 .findFirst()
                 .orElseThrow(() -> new IllegalArgumentException("`principalFactory` '%s' not found.".formatted(principalFactory)))
                 .get();
-    }
-
-    static List<MappingRule> buildMappingRules(@Nullable List<Map> maps) {
-        if (maps == null || maps.isEmpty()) {
-            return List.of(new IdentityMappingRule());
-        }
-        int firstElseIndex = -1;
-        int numElses = 0;
-        for (int i = 0; i < maps.size(); i++) {
-            Map m = maps.get(i);
-
-            if (m.else_() != null) {
-                numElses++;
-                if (firstElseIndex == -1) {
-                    firstElseIndex = i;
-                }
-            }
-        }
-        if (numElses > 1 || (firstElseIndex != -1 && firstElseIndex < maps.size() - 1)) {
-            throw new IllegalArgumentException("An `else` mapping may only occur as the last element of `map`.");
-        }
-        return maps.stream().map(map -> {
-            if (map.replaceMatch() != null) {
-                return new ReplaceMatchMappingRule(map.replaceMatch());
-            }
-            else if (ELSE_IDENTITY.equals(map.else_())) {
-                return new IdentityMappingRule();
-            }
-            else if (ELSE_ANONYMOUS.equals(map.else_())) {
-                return (MappingRule) s -> Optional.empty();
-            }
-            else {
-                throw new IllegalArgumentException("Unknown `else` map '%s', supported values are: '%s', '%s'."
-                        .formatted(map.else_(), ELSE_IDENTITY, ELSE_ANONYMOUS));
-            }
-        }).toList();
     }
 
     @NonNull
